observability/ai_agent/requirements.txt
# AI Incident Insight Agent Requirements

# Core dependencies
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
pyyaml>=6.0
requests>=2.28.0
urllib3>=1.26.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Monitoring and observability
prometheus-client>=0.16.0

# Web framework (for API endpoints)
fastapi>=0.95.0
uvicorn>=0.21.0

# Authentication and security
python-jose>=3.3.0
cryptography>=40.0.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.2.0

# Optional: Machine learning for advanced analysis
scikit-learn>=1.2.0

# Optional: Natural language processing
transformers>=4.25.0
torch>=1.13.0

# Optional: Advanced monitoring
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-prometheus>=1.12.0
